pyproject.toml
[tool.poetry]
name = "llm-showcase"
version = "0.1"
description = "Showcasing LLM capabilities on BBC News"
authors = ["Philippe Büdinger <philippe@buedinger-ai.de>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
bs4 = "^0.0.2"
ipykernel = "^6.29.5"
pandas = "^2.2.2"
nbconvert = "^7.16.4"
jupyter = "^1.0.0"
streamlit = "^1.37.1"
pinecone-client = "^5.0.1"
sentence-transformers = "^3.0.1"
torch = "^2.4.0"
torchvision = "^0.19.0"
torchaudio = "^2.4.0"
openai = "^1.40.6"
python-dotenv = "^1.0.1"
numpy = "1.26.4"
langchain = "^0.2.16"
langchain-community = "^0.2.16"
langchain-openai = "^0.1.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false
